requirements.txt
antlr4-python3-runtime==4.9.2
click==8.0.1
colorama==0.4.4
Flask==2.0.1
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
tk==0.1.0
Werkzeug==2.0.1
